Create interactive-dictionary
In this project I created an interactive python dictionary that reads a json file named data.jason which is a dictionary type file which has words and definitions. All the code is commented for understand-ability and follow-ability. The program prompts the user for a word to look up and through conditionals. If the word is misspelled the program suggests a word that was meant though an imported 3rd party library/module and promoted again if that was what the user meant to look up. if so it will return the suggested word and definition if rejected it will return that it couldn't find the word meant. This was a fun interactive user friendly program to showcase recently learned  python  programming skills.